Add middleware to process multiple tokens concurrently (#4)
* Add middleware to process multiple tokens concurrently

This is done through a new middleware structure called
MultiTokenMiddleware, which will concurrently validate tokens against
mulitple endpoints. These endpoints are meant to be generic, but an
initial implementation is the JWT middleware that we already have in
this repo.

* Reflect error contacting remote endpoint appropraitely

We used to reflect a decoding error if we couldn't contact the remote
JWKS endpoint. Now we give a more appropraite error which makes it easy
for users of this library to debug

* Guard against adding invalid middleware

We were accepting anything that matches the GenericMiddleware interface
in the MultiTokenMiddleware. However... `nil` also matches. So let's
guard against that.

* Refactor: Move generic middleware and multitoken middleware to own package

This creates a clearer separation in the code base.

* Add generic metadata setting function to middleware interface

This allows for any middleware to set the needed metadata in the gin
Context. This is useful for setting logging entries.

* Add implementation of remote middleware

This is a sample implementation that queries a remote endpoint to do
authorization decisions