Respect passed-in Viper instance in helper utility
When calling `RegisterViperOIDCFlags`, the `Viper` instance wasn't being
respected. Instead the underlying function used the global instance...
which is fine in most cases, but is not what was requested by the
library user.

This instead fixes that by actually using a function to bind the flag to
a viper instance using the given instance. The previously used bind
function is kept to ensure backwards compatibility with applications
that might still be using it.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>