add the PublisherSubjectPrefix (#117)
* add the PublisherSubjectPrefix

The NatsJetstream object can be configured with a subject-prefix to
disambiguate subjects between participants on a NATS server. Without
this change, the prefix will be ignored on a publish, resulting in
rejected (or worse, misdirected) messages.

* make description of publishing subject explicit

A consumer of NatsJetstream may not have visibility into how the
connection to NATS was configured (and likely isn't interested). This
change tries to make that relationship more explicit, indicating that
the subject provided by the caller of Publish may only represent a
fraction of the subject to which the message is actually published.