preserve NATS error type
Clients of events (in this case specifically Flasher) rely on the error
type returned from this library for error handling. This change wraps
the NATS error with a library message instead of using the NATS error to
enrich a stock library error. This facilitates a test of `errors.Is(err,
nats.ErrTimeout)` for the case where we want to ignore an error.

An alternative could be to detect the NATS error and rewrite it
conditionally into native events errors, but that is additional work for
no additional operation benefit.